docs: indicate GitHub on license
Note that GitHub is a copyright holder of deno_actions_core because it's
a derivative work of https://github.com/actions/toolkit. Up to this
point, deno_actions_core has been compliant with GitHub's licensing
terms by specifically regarding them in the README. Still, it's good
form to explicitly indicate this in the license as well, hence this
commit.